requirements.txt
flask==3.0.3
flask-login==0.6.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flask-wtf==1.2.1
psycopg2-binary==2.9.9
Pillow==10.3.0
python-dotenv==1.0.1
werkzeug==3.0.3
sqlalchemy==2.0.30
wtforms==3.1.2
gunicorn==22.0.0
rawpy==0.21.0
imageio==2.34.1
boto3==1.34.100
